ena_ethdev: don't override the user provided queue length value
There is a funny logic that seems to be outdated which tries to detect
a situation when a user requests a default size of the queue and configure
a device specific default value.

This seems to be not in line with the DPDK current design that expects
the user to go via the rte_eth_dev_info_get() in order to get device
specific values.

And if the user doesn't care it should pass 0 as a queue length and this
case is also being taken care of on the rte_eth level.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>